Add interface to remove mounts.
We want to add a mount point to /dev/shm using the default spec,
but we don't want to mount /dev/shm via a tmpfs and then add our mountpoint
over it.

RemoveMount will remove a mount point based on the Destination, if it exists.
Does not return any error if the mount point does not exist.  Also added
Mounts() interface so that I could populate a test for new feature. Returns
the list of mounts.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>