runtimetest: improve logic for checking for file modes
Instead of calling `testFileReadAccess()` for all file types, we should
strictly check for file types, to return `errAccess` for other file
types. This error will be skipped during error checks in
`validateMaskedPaths()` and `validateReadonlyPaths()`, so that further
tests can be done even when a single test failed.

Suggested by @wking

Signed-off-by: Dongsu Park <dongsu@kinvolk.io>