validation: add a new test for NSPathMatchTypeError
`checkNSPathMatchType` checks if the container returns an error when
deliberately setting a wrong namespace type. Doing that, it is possible
to verify `NSPathMatchTypeError`, i.e. `The runtime MUST generate an
error if path is not associated with a namespace of type "type"`.

See also https://github.com/opencontainers/runtime-tools/issues/572

Signed-off-by: Dongsu Park <dongsu@kinvolk.io>