AddDevice(): better diagnostic when creating dup
The existing diagnostic[1] appeared in a CI job with no
possible hope of figuring out the source of the problem.
This PR amends the diagnostic to include the name of
the requested newly-created device as well as the
existing "duplicate".

 [1] WARNING: The same type, major and minor should not be used for multiple devices.

(I am not convinced that this condition merits a warning
but will leave that discussion to better minds than mine).

Signed-off-by: Ed Santiago <santiago@redhat.com>