migrate to go modules
This switches the repository to use go modules. This commit is a "mechanical"
migration, keeping all dependencies to the same version as is currently used,
and is the result of:

    go mod init
    go mod tidy

After that, I updated the go version in go.mod to 1.11, which is the oldest
version currently tested against (and first go release with go modules, although
go modules should not be required, as vendoring is used).

It looks like `go mod vendor` ignores LICENSE files with an extension, as a result,
the license for testify is removed in this commit;
vendor/github.com/stretchr/testify/LICENCE.txt

Another LICENSE is removed in gojsonschema, which is correct, as the license
covered files that were not vendored;
vendor/github.com/xeipuuv/gojsonschema/json_schema_test_suite/LICENSE

The testify license is "to be discussed"; license with an extension is non-standard,
but may occur; the issue will be resolved when updating to a more recent version
of testify; the version currently used is very old (pre v1.2.0, Aug 10, 2017);
https://github.com/stretchr/testify/tree/890a5c3458b43e6104ff5da8dfa139d013d77544),
and current version have renamed the file to LICENSE.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>