Makefile: use fancy git commit ids
Now, the commit ID will contain the closest tag, the number of commits
since the tag, and the (abbreviated) git commit sha (see example below).

This tag is still unique and can be used instead of bare sha  for all
git commands.

Before:
	$ ./runtimetest -version
	runtimetest version 0.9.0, commit: ea2948e9efbfc634991aad90ebb6bf54995c9a37

After:
	$ ./runtimetest -version
	runtimetest version 0.9.0, commit: v0.9.0-28-ge91e387

This means that
 - the closest tag is v0.9.0
 - there were 28 commits after the tag
 - the abbreviated sha is e91e387

Nice, isn't it?

While at it,
 - allow to set commit id via environment (might be needed for distros);
 - remove unneeded "|| true" stance.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>