cmd/runtimetest: use github.com/moby/sys/mountinfo
I noticed that this package was a fork of an old version of the "mount" package
in the docker repository. This package was migrated to a separate module in the
github.com/moby/sys repository, which should be easy to consume now that it's
no longer insxide the docker repository.

This patch switches the local fork to the new module.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>