[0.9.1] spec: update to runtime-spec 1.0.2
Update our spec version to a properly released spec, in preparation for
the 0.9.1 release.

NOTE: This commit only exists in the release-0.9.1 branch of
runtime-tools. There is no real corresponding commit in main, as we have
historically not tracked the runtime-spec per-release. Commit
43243fed9790 ("Add missing interface to set init processes Umask") did
update main to 1.0.2 but there were other patches necessary to support
all of the spec features in 1.0.2.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>